Move parquet_column to lib.rs
