+++ parquet/src/arrow/arrow_reader/statistics.rs
@@ -18,6 +18,7 @@
 //! [`StatisticsConverter`] to convert statistics in parquet format to arrow [`ArrayRef`].
 
 use crate::arrow::buffer::bit_util::sign_extend_be;
+use crate::arrow::parquet_column;
 use crate::data_type::{ByteArray, FixedLenByteArray};
 use crate::errors::{ParquetError, Result};
 use crate::file::metadata::{ParquetColumnIndex, ParquetOffsetIndex, RowGroupMetaData};
@@ -36,7 +37,7 @@ use arrow_array::{
     UInt32Array, UInt64Array, UInt8Array,
 };
 use arrow_buffer::i256;
-use arrow_schema::{DataType, Field, FieldRef, Schema, TimeUnit};
+use arrow_schema::{DataType, Field, Schema, TimeUnit};
 use half::f16;
 use paste::paste;
 use std::sync::Arc;
@@ -947,32 +948,6 @@ macro_rules! get_data_page_statistics {
         }
     }
 }
-
-/// Lookups up the parquet column by name
-///
-/// Returns the parquet column index and the corresponding arrow field
-pub fn parquet_column<'a>(
-    parquet_schema: &SchemaDescriptor,
-    arrow_schema: &'a Schema,
-    name: &str,
-) -> Option<(usize, &'a FieldRef)> {
-    let (root_idx, field) = arrow_schema.fields.find(name)?;
-    if field.data_type().is_nested() {
-        // Nested fields are not supported and require non-trivial logic
-        // to correctly walk the parquet schema accounting for the
-        // logical type rules - <https://github.com/apache/parquet-format/blob/master/LogicalTypes.md>
-        //
-        // For example a ListArray could correspond to anything from 1 to 3 levels
-        // in the parquet schema
-        return None;
-    }
-
-    // This could be made more efficient (#TBD)
-    let parquet_idx = (0..parquet_schema.columns().len())
-        .find(|x| parquet_schema.get_column_root_idx(*x) == root_idx)?;
-    Some((parquet_idx, field))
-}
-
 /// Extracts the min statistics from an iterator of [`ParquetStatistics`] to an
 /// [`ArrayRef`]
 ///
